Scenario
Having scraped and shared the specific data required by the executives, claims department, and clients of your organisation, you now need to share it with them. You know that none of the stakeholders involved is as tech-savvy or data-savvy as you are and that you need to share the data with them in a way they can understand and use. To this end, you’ve been asked to download your scraped data as a CSV file. 

Additionally, other data analysts at your organisation want access to your data and have requested the data also be shared as JSON files. Your analytics department has also asked you to focus on the six continents on the table, and identify any missing data.